+++ lib/librarian/puppet/source.rb
@@ -1,4 +1,3 @@
-require 'librarian/puppet/util'
 require 'librarian/puppet/requirement'
 require 'librarian/puppet/source/path'
 require 'librarian/puppet/source/git'